added paging tool helper and extracted a lot of math somewhere else
