+++ src/web/components/ClientSideTokenGeneration/CstgDomainsTable.tsx
@@ -2,7 +2,8 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { CheckedState } from '@radix-ui/react-checkbox';
 import { useEffect, useState } from 'react';
 
-import { PagingTool, RowsPerPageValues } from '../Core/PagingTool';
+import { PagingTool } from '../Core/PagingTool';
+import { RowsPerPageValues } from '../Core/PagingToolHelper';
 import { TableNoDataPlaceholder } from '../Core/TableNoDataPlaceholder';
 import { TriStateCheckbox, TriStateCheckboxState } from '../Core/TriStateCheckbox';
 import CstgAddDomainDialog from './CstgAddDomainDialog';
